5_New_Tech/AES_CTR_Demo/python_implementation/requirements.txt
pycryptodome>=3.9 